Moving <Filter> definition file around.
