+++ app/javascripts/components/Schedule/filter.css
