+++ app/javascripts/components/Schedule/filter.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./styles.css";
+import styles from "./filter.css";
 
 export default React.createClass({
   displayName: "Filter",
